+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Upgraded url_launcher plugin to resolve build issue on Xcode 16.
 - Survey Card widgets color explicitly set to be white [#4385](https://github.com/rokwire/illinois-app/issues/4385).
+### Fixed
+- Link / Unlink event to group [#481](https://github.com/rokwire/app-flutter-plugin/issues/481).
 
 ## [1.7.2] - 2024-09-20
 ### Fixed