+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Survey Card widgets color explicitly set to be white [#4385](https://github.com/rokwire/illinois-app/issues/4385).
 ### Fixed
 - Link / Unlink event to group [#481](https://github.com/rokwire/app-flutter-plugin/issues/481).
+### Added
+- Added Places model and service [#475](https://github.com/rokwire/app-flutter-plugin/issues/475).
 
 ## [1.7.2] - 2024-09-20
 ### Fixed